feat: changed function for getting Grafana URLs to get all load balancers and read the tags that we're sending
